requirements.txt
discord.py
aiohttp
pytz
